Issue number:  MOD-23 Obtained from: JIRA Submitted by:  Steven Devijver Reviewed by:  -
Add support to retrieve the arguments for the IN and NOT IN operators from a bean property. The value should be a collection, an iterator, an enumeration or an object array. Only one bean property can be specified and must be pre-fixed with "@".

Example:

{ size : ? in @sizes : 'Size is not correct.' }

Issue number:  MOD-24
Obtained from: JIRA
Submitted by:  Steven Devijver
Reviewed by:  -

Support retrieval of a value from a map. The map is provided by a bean property. The syntax is : "$" <PATH> ( function() )

Example:

{ firstName : ? = $parameters('firstName') : 'FirstName is not correct.' }